Bug fixes and small features
- Fixed mouse problems
- Added Level selection state
- Made wave difficulty increase naturally

+++ src/core/Entity.java
@@ -14,7 +14,7 @@ public abstract class Entity extends DrawableImage {
     public float speed = 0.4f;
     public boolean alive = true;
     public GameContainer c;
-    public int expOutput = 10;
+    public float expOutput = 10;
     public float delta;
     
     /**
